Move tests to submodule
